# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Garantia Construtora",
        "version": "1.0"
    },
    "paths": {
...
        "/api/logs/": {
            "get": {
                "tags": [
                    "logs"
                ],
                "summary": "Show logs on system",
                "description": "On this route, we can set '*-order' with 'asc' or 'desc'\n    *          and '*-search' with any word, in '*', we can too set specifics DB column to\n    *          compare between dates in '*-search' we need set a pipe separing the\n    *          values, like 'date1|date2', pipe can be used to set more search\n    *          words too, but remember, they will compare using 'AND' in 'WHERE'\n    *          clasule, they dynamic values only can't be setted on SwaggerUI.",
                "operationId": "logs.index",
                "parameters": [
                    {
                        "name": "business",
                        "in": "query",
                        "description": "Business's ID, if used a user with management role, don't set it",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Rows limit by page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show logs available"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized (User don't have permission, access token expired or isn't logged yet)"
                    },
                    "500": {
                        "description": "API internal error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/logs/export": {
            "get": {
                "tags": [
                    "logs"
                ],
                "summary": "Export logs on system to a file",
                "description": "On this route, we can set '*-order' with 'asc' or 'desc'\n    *          and '*-search' with any word, in '*', we can too set specifics DB column to\n    *          compare between dates in '*-search' we need set a pipe separing the\n    *          values, like 'date1|date2', pipe can be used to set more search\n    *          words too, but remember, they will compare using 'AND' in 'WHERE'\n    *          clasule, they dynamic values only can't be setted on SwaggerUI.",
                "operationId": "logs.export",
                "parameters": [
                    {
                        "name": "business",
                        "in": "query",
                        "description": "Business's ID, if used a user with management role, don't set it",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "building",
                        "in": "query",
                        "description": "Building's ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Rows limit by page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a link to download exported file"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized (User don't have permission, access token expired or isn't logged yet)"
                    },
                    "500": {
                        "description": "API internal error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/roles/available": {
            "get": {
                "tags": [
                    "roles"
                ],
                "summary": "Show available roles to user attach",
                "description": "<b>Important:</b><br>\n    *          Business's ID need to be setted if user authenticated:<br>\n    *          1 - Don't have a management role with permission 'role > read' enabled or;<br>\n    *          2 - Have a role with permission 'role > read' enabled in a specific business;<br>\n    *          On first case, we can see any role, but in second case we can only see info of roles how\n    *          is attached on that business.",
                "operationId": "roles.show.available",
                "parameters": [
                    {
                        "name": "business",
                        "in": "query",
                        "description": "Business's ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show available roles to use"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized (User don't have permission, access token expired or isn't logged yet)"
                    },
                    "500": {
                        "description": "API internal error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}